Enhance code readability
